+++ hw/dv/verilator/memutil_dpi.core
@@ -1,0 +1,20 @@
+CAPI=2:
+# Copyright lowRISC contributors.
+# Licensed under the Apache License, Version 2.0, see LICENSE for details.
+# SPDX-License-Identifier: Apache-2.0
+
+name: "lowrisc:dv_verilator:memutil_dpi"
+description: "DPI memory utilities"
+filesets:
+  files_cpp:
+    files:
+      - cpp/dpi_memutil.cc
+      - cpp/dpi_memutil.h: { is_include_file: true }
+      - cpp/sv_scoped.cc
+      - cpp/sv_scoped.h: { is_include_file: true }
+    file_type: cppSource
+
+targets:
+  default:
+    filesets:
+      - files_cpp
